requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
python-docx==0.8.11
pdfplumber==0.7.6
pandas==2.2.3
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.1
PyPDF2==3.0.1
Werkzeug==2.2.3
